fix(gametime): drop a no-op uint64 conversion CI lint flagged
golangci-lint failed the PR on one issue:
internal/gametime/gametime.go:638:24: unnecessary conversion (unconvert).

roundsPerDay is declared uint64 at line 572, so uint64(roundsPerDay) was
always a no-op. The line is PRE-EXISTING and predates this branch. It surfaced
now because Task 5 removed the redundant float64 wraps on the neighbouring
nightHoursPerDay expressions, gofmt realigned the trailing comments across the
whole block, and the line therefore entered the diff as a changed line. CI runs
golangci-lint with --new-from-patch, which classifies issues by whether their
line is in the patch rather than by whether the defect is new, so a
whitespace-only realignment pulled an old issue into scope.

Worth remembering: a gofmt realignment can drag pre-existing lint into a PR's
"new issues" set, and unconvert is not part of go vet or gofmt, so no local
gate on this branch could have caught it. golangci-lint 2.12.2 is installed
locally and `golangci-lint run --new-from-merge-base=origin/master` reproduces
CI's check exactly; it now reports 0 issues.

CI's test job was green throughout. Only lint was red.

+++ internal/gametime/gametime.go
@@ -635,7 +635,7 @@ func GetLastPeriod(periodName string, roundNumber uint64) uint64 {
 	} else if periodName == `sunrise` { // last sunrise
 
 		roundNumber -= roundOfDay                                              // Strip rounds of today off
-		roundNumber -= uint64(roundsPerDay)                                    // Subtract a day
+		roundNumber -= roundsPerDay                                            // Subtract a day; roundsPerDay is already uint64
 		roundNumber += uint64(math.Ceil(nightHoursPerDay / 2 * roundsPerHour)) // add half a night
 
 	} else if periodName == `sunset` { // 12am of next day, minus half of night
